+++ example/lib/playlist_page/playlist_page.dart
@@ -31,10 +31,11 @@ class _PlaylistPageState extends State<PlaylistPage> {
     dataSourceList.add(BetterPlayerDataSource(
         BetterPlayerDataSourceType.NETWORK,
         "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4"));
-    dataSourceList.add(BetterPlayerDataSource(
-        BetterPlayerDataSourceType.NETWORK,
-        "http://sample.vodobox.com/skate_phantom_flex_4k/skate_phantom_flex_4k.m3u8",
-        liveStream: true));
+    dataSourceList.add(
+      BetterPlayerDataSource(BetterPlayerDataSourceType.NETWORK,
+          "http://sample.vodobox.com/skate_phantom_flex_4k/skate_phantom_flex_4k.m3u8",
+          liveStream: true),
+    );
 
     return dataSourceList;
   }
@@ -67,6 +68,7 @@ class _PlaylistPageState extends State<PlaylistPage> {
                 betterPlayerConfiguration: BetterPlayerConfiguration(
                     autoPlay: true,
                     aspectRatio: 1,
+                    rotation: 90,
                     fit: BoxFit.cover,
                     subtitlesConfiguration:
                         BetterPlayerSubtitlesConfiguration(fontSize: 10),