+++ lib/src/core/better_player_with_controls.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:math';
 
 import 'package:better_player/src/controls/better_player_controls_configuration.dart';
 import 'package:better_player/src/controls/better_player_cupertino_controls.dart';
@@ -53,8 +54,7 @@ class _BetterPlayerWithControlsState extends State<BetterPlayerWithControls> {
   }
 
   void _onControllerChanged() {
-    setState(() {
-    });
+    setState(() {});
   }
 
   @override
@@ -77,14 +77,24 @@ class _BetterPlayerWithControlsState extends State<BetterPlayerWithControls> {
 
   Container _buildPlayerWithControls(
       BetterPlayerController betterPlayerController, BuildContext context) {
+    var configuration = betterPlayerController.betterPlayerConfiguration;
+    var rotation = configuration.rotation;
+
+    if (!(rotation <= 360 && rotation % 90 == 0)){
+      print("Invalid rotation provided. Using rotation = 0");
+      rotation = 0;
+    }
     return Container(
       child: Stack(
         fit: StackFit.passthrough,
         children: <Widget>[
           betterPlayerController.placeholder ?? Container(),
-          _BetterPlayerVideoFitWidget(
-            betterPlayerController,
-            betterPlayerController.betterPlayerConfiguration.fit,
+          Transform.rotate(
+            angle: rotation * pi / 180,
+            child: _BetterPlayerVideoFitWidget(
+              betterPlayerController,
+              betterPlayerController.betterPlayerConfiguration.fit,
+            ),
           ),
           betterPlayerController.overlay ?? Container(),
           BetterPlayerSubtitlesDrawer(